added comments page

+++ models/posts.js
@@ -1,4 +1,5 @@
 const Post = require('../lib/mongo').Post;
+const CommentModel = require('./comments');
 const marked = require('marked');
 
 Post.plugin('contentToHtml',{
@@ -15,7 +16,24 @@ Post.plugin('contentToHtml',{
     return post;
   }
 });
-
+Post.plugin('addCommentsCount',{
+  afterFind: function(posts){
+    return Promise.all(posts.map(function(post){
+      return CommentModel.getCommentsCount(post._id).then(function(commentsCount){
+        post.commentsCount = commentsCount;
+        return post;
+      })
+    }))
+  },
+  afterFindOne: function(post){
+    if(post){
+      return CommentModel.getCommentsCount(post._id).then(function(commentsCount){
+        post.commentsCount = commentsCount;
+        return post;
+      })
+    }
+  }
+});
 module.exports = {
   create: function create(post){
     return Post.create(post).exec();
@@ -25,6 +43,7 @@ module.exports = {
     .findOne({_id: id})
     .populate({path:'author',model: 'User'})
     .addCreatedAt()
+    .addCommentsCount()
     .contentToHtml()
     .exec();
   },
@@ -37,6 +56,7 @@ module.exports = {
     .populate({path:'author',model:'User'})
     .sort({_id: -1})
     .addCreatedAt()
+    .addCommentsCount()
     .contentToHtml()
     .exec();
   },
@@ -54,6 +74,12 @@ module.exports = {
     return Post.update({_id: postId },{$set: data}).exec();
   },
   delPostById: function delPostById(postId){
-    return Post.deleteOne({_id:postId}).exec();
+    return Post.deleteOne({author: author,_id:postId}).exec()
+      .then(function(res){
+        if(res.result.ok && res.result.n > 0){
+          //接着删除文章下所有留言
+          return CommentModel.delCommentsByPostId(postId);
+        }
+      })
   }
 }
